+++ core/src/main/java/io/questdb/cairo/replication/TableReplicationStreamHeaderSupport.java
@@ -10,32 +10,32 @@ public class TableReplicationStreamHeaderSupport {
     public static final byte FRAME_TYPE_MAX_ID = FRAME_TYPE_COMMIT_BLOCK;
     public static final byte FRAME_TYPE_UNKNOWN = (byte) 0xff;
 
-    public static final long OFFSET_FRAME_SIZE = 0;
-    public static final long OFFSET_FRAME_TYPE = OFFSET_FRAME_SIZE + Integer.BYTES;
-    public static final long OFFSET_MASTER_TABLE_ID = OFFSET_FRAME_TYPE + Byte.BYTES;
+    public static final int OFFSET_FRAME_SIZE = 0;
+    public static final int OFFSET_FRAME_TYPE = OFFSET_FRAME_SIZE + Integer.BYTES;
+    public static final int OFFSET_MASTER_TABLE_ID = OFFSET_FRAME_TYPE + Byte.BYTES;
 
-    public static final long OFFSET_DF_FIRST_TIMESTAMP = OFFSET_MASTER_TABLE_ID + Integer.BYTES;
-    public static final long OFFSET_DF_COLUMN_INDEX = OFFSET_DF_FIRST_TIMESTAMP + Long.BYTES;
-    public static final long OFFSET_DF_DATA_OFFSET = OFFSET_DF_COLUMN_INDEX + Integer.BYTES;
-    public static final long DF_HEADER_SIZE = OFFSET_DF_DATA_OFFSET + Long.BYTES;
+    public static final int OFFSET_DF_FIRST_TIMESTAMP = OFFSET_MASTER_TABLE_ID + Integer.BYTES;
+    public static final int OFFSET_DF_COLUMN_INDEX = OFFSET_DF_FIRST_TIMESTAMP + Long.BYTES;
+    public static final int OFFSET_DF_DATA_OFFSET = OFFSET_DF_COLUMN_INDEX + Integer.BYTES;
+    public static final int DF_HEADER_SIZE = OFFSET_DF_DATA_OFFSET + Long.BYTES;
 
-    public static final long OFFSET_SFF_COLUMN_INDEX = OFFSET_MASTER_TABLE_ID + Integer.BYTES;
-    public static final long OFFSET_SFF_DATA_OFFSET = OFFSET_SFF_COLUMN_INDEX + Integer.BYTES;
-    public static final long SFF_HEADER_SIZE = OFFSET_SFF_DATA_OFFSET + Long.BYTES;
+    public static final int OFFSET_SFF_COLUMN_INDEX = OFFSET_MASTER_TABLE_ID + Integer.BYTES;
+    public static final int OFFSET_SFF_DATA_OFFSET = OFFSET_SFF_COLUMN_INDEX + Integer.BYTES;
+    public static final int SFF_HEADER_SIZE = OFFSET_SFF_DATA_OFFSET + Long.BYTES;
 
-    public static final long SCR_HEADER_SIZE = OFFSET_MASTER_TABLE_ID + Integer.BYTES;
+    public static final int SCR_HEADER_SIZE = OFFSET_MASTER_TABLE_ID + Integer.BYTES;
 
-    public static final long CB_HEADER_SIZE = OFFSET_MASTER_TABLE_ID + Integer.BYTES;
+    public static final int CB_HEADER_SIZE = OFFSET_MASTER_TABLE_ID + Integer.BYTES;
 
-    public static final long OFFSET_EOB_N_FRAMES_SENT = OFFSET_MASTER_TABLE_ID + Integer.BYTES;
-    public static final long EOB_HEADER_SIZE = OFFSET_EOB_N_FRAMES_SENT + Integer.BYTES;
+    public static final int OFFSET_EOB_N_FRAMES_SENT = OFFSET_MASTER_TABLE_ID + Integer.BYTES;
+    public static final int EOB_HEADER_SIZE = OFFSET_EOB_N_FRAMES_SENT + Integer.BYTES;
 
-    public static final long MIN_HEADER_SIZE = CB_HEADER_SIZE;
-    public static final long MAX_HEADER_SIZE = DF_HEADER_SIZE;
+    public static final int MIN_HEADER_SIZE = CB_HEADER_SIZE;
+    public static final int MAX_HEADER_SIZE = DF_HEADER_SIZE;
 
-    private static long[] HEADER_SIZE_BY_FRAME_TYPE;
+    private static int[] HEADER_SIZE_BY_FRAME_TYPE;
     static {
-        HEADER_SIZE_BY_FRAME_TYPE = new long[FRAME_TYPE_MAX_ID - FRAME_TYPE_MIN_ID + 1];
+        HEADER_SIZE_BY_FRAME_TYPE = new int[FRAME_TYPE_MAX_ID - FRAME_TYPE_MIN_ID + 1];
         HEADER_SIZE_BY_FRAME_TYPE[FRAME_TYPE_DATA_FRAME - FRAME_TYPE_MIN_ID] = DF_HEADER_SIZE;
         HEADER_SIZE_BY_FRAME_TYPE[FRAME_TYPE_SYMBOL_STRINGS_FRAME - FRAME_TYPE_MIN_ID] = SFF_HEADER_SIZE;
         HEADER_SIZE_BY_FRAME_TYPE[FRAME_TYPE_END_OF_BLOCK - FRAME_TYPE_MIN_ID] = EOB_HEADER_SIZE;
@@ -43,7 +43,7 @@ public class TableReplicationStreamHeaderSupport {
         HEADER_SIZE_BY_FRAME_TYPE[FRAME_TYPE_COMMIT_BLOCK - FRAME_TYPE_MIN_ID] = CB_HEADER_SIZE;
     }
 
-    public static final long getFrameHeaderSize(byte frameType) {
+    public static final int getFrameHeaderSize(byte frameType) {
         return HEADER_SIZE_BY_FRAME_TYPE[frameType - FRAME_TYPE_MIN_ID];
     }
 }
